+++ Resources/Locale/en-US/silicons/derelict/role.ftl
@@ -1,4 +1,4 @@
 derelict-cyborg-round-end-agent-name = derelict cyborg
 
 derelict-cyborg-role-greeting =
-    You are a cyborg that has been lost in space for many years that has now drifted close to a space station. You can use your fire extinguisher and GPS to get board the station. Remember to follow your laws. #Greeting is unused for now.
+    You are a cyborg that has been lost in space for many years that has now drifted close to a space station. You can use your fire extinguisher and GPS to get board the station. Remember to follow your laws.
